Add quiz.js
Add quiz.js that provides the functionality to attempt quizzes. It ensures all questions have been answered, determines whether questions have been answered correctly or not, and outputs the results to the user.

Add a procedure that includes all code for copying the necessary JavaScript to the build folder.

Tidy up testquiz.txt.