Modify build procedure
Quizzes are now output to directories named according to their title. Existing directories are replaced when re-exporting.

The program no longer asks the user if they would like to remove the build folder or not.

Update README.md to reflect changes.

These changes simplify the build process and makes it easier to export multiple quizzes in sequence.